LWS_PRE: make int

+++ lib/libwebsockets.h
@@ -4011,7 +4011,7 @@ lws_set_timeout(struct lws *wsi, enum pending_timeout reason, int secs);
 */
 //@{
 #if !defined(LWS_SIZEOFPTR)
-#define LWS_SIZEOFPTR (sizeof (void *))
+#define LWS_SIZEOFPTR ((int)sizeof (void *))
 #endif
 #if !defined(u_int64_t)
 #define u_int64_t unsigned long long
